cached assets are now organized into folders better behaviour for otfimage and otfimagereceiver, added otfimagereceiver::getImageNum otfimage::atSize($params) : $params can now be either a string (var1:val1,var2:val2....) or an array
git-svn-id: https://m4php5.googlecode.com/svn/trunk@236 4cff5740-f916-11dd-b507-8b4ecd5210be

+++ commands/clearcache.php
@@ -47,7 +47,7 @@ class Command_clearcache extends Command
   }
   public function clearAssetsCache()
   {
-      $this->_emptyfolder(APP_ROOT.WEB_FOLDER.'/cache',false);
+      $this->_emptyfolder(APP_ROOT.WEB_FOLDER.'/cache',true);
   }
   public function clearConfigCache()
   {
@@ -68,9 +68,12 @@ class Command_clearcache extends Command
     if(is_dir($cacheFolder)) {
       $nbfiles = 0;
       foreach(FileUtils::getAllFiles($cacheFolder,'',false) as $file) {
-        if(is_file($file)) {
+        if(is_file($file) && !$recursive) {
           unlink($file);
           $nbfiles++;
+        } elseif(is_dir($file)) {
+          $this->_emptyfolder($file,true);
+          @rmdir($file);
         }
       }
       $this->line('Clearing dir '.$cacheFolder);
